requirements.txt
openai
py3-tts
python-dotenv
SpeechRecognition
PyAudio
pocketsphinx
